feat(component): card
- Add card component and story
- Fix broken file path for FloatingButton
- Fix linting issues
