--- src/components/Card/Card.stories.ts
@@ -1,18 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { Card } from './Card';
-
-const meta: Meta<typeof Card> = {
-  component: Card,
-  title: 'Components/Data Display/Card',
-  tags: ['autodocs']
-};
-export default meta;
-
-type Story = StoryObj<typeof Card>;
-
-export const Default: Story = {
-  args: {
-    children: 'Card'
-  }
-};
